+++ packages/runtime/plugin-testing/tests/bff/utils.test.ts
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment node
+ */
 import * as path from 'path';
 import { isBFFProject } from '../../src/cli/bff/utils';
 import mockAPI from '../../src/cli/bff/mockAPI';