+++ packages/server/core/src/index.ts
@@ -1,4 +1,5 @@
-export * from './plugin';
+export * from './core/plugin';
+export * from './core/render';
 export * from '@modern-js/plugin';
-export * from './loadPlugins';
+export * from './core/loadPlugins';
 export * from './types/config';