+++ packages/server/server/src/index.ts
@@ -1,22 +1,5 @@
-import {
-  DevServer as Server,
-  DevServerForRsbuild as ServerForRsbuild,
-} from './server';
-import type {
+export { createDevServer } from './createDevServer';
+export type {
   ModernDevServerOptions,
-  ModernDevServerOptionsNew,
-} from './types';
-
-export { Server, ServerForRsbuild };
-export type { ModernDevServerOptions, ModernDevServerOptionsNew };
-
-// TODO: it seems not used in any pkgs?
-export default (options: ModernDevServerOptions): Promise<Server> => {
-  if (options == null) {
-    throw new Error('can not start server without options');
-  }
-
-  const server = new Server(options);
-
-  return server.init();
-};
+  InitProdMiddlewares,
+} from './createDevServer';