+++ packages/solutions/app-tools/src/commands/dev.ts
@@ -1,11 +1,10 @@
 import { PluginAPI, ResolvedConfigContext } from '@modern-js/core';
 import { DEFAULT_DEV_HOST } from '@modern-js/utils';
+import { createDevServer } from '@modern-js/server';
+import { initProdMiddlewares } from '@modern-js/prod-server';
+import { registerCompiler } from '../utils/register';
 import { printInstructions } from '../utils/printInstructions';
-import {
-  setServer,
-  createServer,
-  injectDataLoaderPlugin,
-} from '../utils/createServer';
+import { setServer } from '../utils/createServer';
 import { generateRoutes } from '../utils/routes';
 import { DevOptions } from '../utils/types';
 import { buildServerConfig } from '../utils/config';
@@ -29,6 +28,12 @@ export const dev = async (
   const appContext = api.useAppContext();
   const hookRunners = api.useHookRunners();
 
+  await registerCompiler(
+    appContext.appDirectory,
+    appContext.distDirectory,
+    normalizedConfig?.source?.alias,
+  );
+
   normalizedConfig = { ...normalizedConfig, cliOptions: options };
   ResolvedConfigContext.set(normalizedConfig);
 
@@ -39,6 +44,7 @@ export const dev = async (
     apiOnly,
     serverConfigFile,
     metaName,
+    serverRoutes,
   } = appContext;
 
   await buildServerConfig({
@@ -60,6 +66,7 @@ export const dev = async (
   const serverInternalPlugins = await getServerInternalPlugins(api);
 
   const serverOptions = {
+    metaName,
     dev: {
       port,
       https: normalizedConfig.dev.https,
@@ -67,24 +74,24 @@ export const dev = async (
       ...normalizedConfig.tools?.devServer,
     },
     appContext: {
-      metaName,
-      appDirectory: appContext.appDirectory,
-      sharedDirectory: appContext.sharedDirectory,
+      appDirectory,
       apiDirectory: appContext.apiDirectory,
       lambdaDirectory: appContext.lambdaDirectory,
+      sharedDirectory: appContext.sharedDirectory,
     },
+    routes: serverRoutes,
     pwd: appDirectory,
-    config: normalizedConfig,
+    config: normalizedConfig as any,
     serverConfigFile,
-    internalPlugins: injectDataLoaderPlugin(serverInternalPlugins),
+    internalPlugins: serverInternalPlugins,
     ...devServerOptions,
   };
 
   if (apiOnly) {
-    const app = await createServer({
-      ...(serverOptions as any),
-      compiler: null,
-    });
+    const app = await createDevServer(
+      serverOptions as any,
+      initProdMiddlewares,
+    );
 
     const host = normalizedConfig.dev?.host || DEFAULT_DEV_HOST;
 
@@ -93,17 +100,16 @@ export const dev = async (
         port,
         host,
       },
-      async (err: Error) => {
-        if (err) {
-          throw err;
-        }
+      () => {
         printInstructions(hookRunners, appContext, normalizedConfig);
       },
     );
   } else {
     const { server } = await appContext.builder!.startDevServer({
-      serverOptions: serverOptions as any,
+      serverOptions,
+      initProdMiddlewares,
     });
-    setServer(server);
+    // TODO: set correct server
+    setServer(server as any);
   }
 };