+++ tests/integration/ssr/fixtures/base/src/routes/layout.tsx
@@ -1,7 +1,8 @@
-import { Link, Outlet } from '@modern-js/runtime/router';
+import { Link, Outlet, useLoaderData } from '@modern-js/runtime/router';
 
-let count = 0;
 export default function Layout() {
+  const { count } = useLoaderData() as { count: number };
+
   return (
     <div>
       Root layout
@@ -20,7 +21,7 @@ export default function Layout() {
         </Link>
       </div>
       <Outlet />
-      <footer>count:{count++}</footer>
+      <footer>{`count:${count}`}</footer>
     </div>
   );
 }