+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-changeset
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/plugin-i18n@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes