+++ packages/cli/plugin-proxy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-proxy
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes