+++ packages/runtime/plugin-testing/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/plugin-testing
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/runtime@2.48.5
+  - @modern-js/bff-core@2.48.5
+  - @modern-js/prod-server@2.48.5
+  - @modern-js/babel-compiler@2.48.5
+  - @modern-js/plugin@2.48.5
+  - @modern-js/types@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes