+++ packages/solutions/app-tools/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/app-tools
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/core@2.48.5
+  - @modern-js/plugin-data-loader@2.48.5
+  - @modern-js/plugin-i18n@2.48.5
+  - @modern-js/plugin-lint@2.48.5
+  - @modern-js/uni-builder@2.48.5
+  - @modern-js/server-core@2.48.5
+  - @modern-js/prod-server@2.48.5
+  - @modern-js/server@2.48.5
+  - @modern-js/server-utils@2.48.5
+  - @modern-js/node-bundle-require@2.48.5
+  - @modern-js/plugin@2.48.5
+  - @modern-js/types@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes