+++ packages/solutions/monorepo-tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/monorepo-tools
 
+## 2.48.5
+
+### Patch Changes
+
+- Updated dependencies [4ca9f4c]
+  - @modern-js/utils@2.48.5
+  - @modern-js/core@2.48.5
+  - @modern-js/plugin-changeset@2.48.5
+  - @modern-js/plugin-i18n@2.48.5
+  - @modern-js/plugin-lint@2.48.5
+  - @modern-js/plugin@2.48.5
+
 ## 2.48.4
 
 ### Patch Changes